refactor(dashboard): split log drain views by sink
Amp-Thread-ID: https://ampcode.com/threads/T-01a03c72-c96e-75fb-9dc1-afdd7d1f0687
