+++ web/apps/dashboard/app/(app)/[workspaceSlug]/settings/logdrains/[drainId]/drain-detail.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { match } from "@unkey/match";
+import { AxiomDrainDetail } from "./axiom-drain-detail";
+import { HttpDrainDetail } from "./http-drain-detail";
+import type { Drain, DrainTelemetry } from "./types";
+
+export function DrainDetail({ drain, ...telemetry }: { drain: Drain } & DrainTelemetry) {
+  return match(drain)
+    .with({ kind: "http" }, (httpDrain) => (
+      <HttpDrainDetail key={httpDrain.id} drain={httpDrain} {...telemetry} />
+    ))
+    .with({ kind: "axiom" }, (axiomDrain) => (
+      <AxiomDrainDetail key={axiomDrain.id} drain={axiomDrain} {...telemetry} />
+    ))
+    .exhaustive();
+}
